requirements.txt
fastapi==0.95.2
pydantic==1.10.13
uvicorn==0.20.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.3
httpx==0.24.1
python-jose[cryptography]==3.3.0